Maintain attributes in a sub-object instead of on the element
The reflection of properties to attributes using the element instance as their owning object cleverly needs only a little implementation. But it has some side-effects, causing the element behave differently than in the browser DOM:

* el.setAttribute("foo", "bar") creates a property "foo" on the element
* a property el.foo = "bar" will be replaced by el.setAttribute("foo", "bar")

Because this behaviour is not what a usual component would depend on, a conflict with an actual implementation would be rare.

Another problem could arise from requiring the attributes be represented only as owned properties of the element instance. It will prevent testing those custom element libraries, which use class inheritance together with default property values.

For example, the "type" will not be usable as an attribute of HTMLPasswordElement, because it is used in its ancestor class already and thus not owned by the final instance:

    class HTMLInputElement extends HTMLElement {
      type = "text"
    }

    class HTMLPasswordElement extends HTMLInputElement {
      type = "password"
    }

For example, the "count" will not be usable as an attribute of TestComponent, because the class decorator is enhancing the class by inheritance:

    @comp({ tag: 'test-component' })
    export class TestComponent extends HTMLElement {
      @prop({ type: 'number' })
      count = 0
    }

    function comp(ctor) {
      class Comp extends ctor {
        ...
      }
      customElements.define(tag, Comp)
      return Comp
    }

Yet another problem will arise if the property decorator will create a property with get and set accessors, whih will add it to the prototype:

    function prop({ type }) {
      return (proto, name): {} => {
        ...
        return {
          enumerable: true,
          get() { return ... },
          set(value: any) { ... }
        }
      }
    }

The workaround by listing properties that should reflect to attributes for all element types is naive, but should work as a simplification for testing purposes, because properties of specialised elements like input.value are not usually accessed on general elements like div.

WARNING: The style attribute is always rendered at the end. This may break existing tests that depend on the order that the attributes are rendered to outerHTML.